backend/requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.1
SQLAlchemy==2.0.21
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
